requirements.txt
SpeechRecognition
pyttsx3
pyaudio
Pillow
